Require macOS 13
MCP SDK requires macOS 13 and up.
In SPM platforms are declared at the package level, not the target level, meaning we need to drop earlier versions on all targets even though this shouldn't be a requirement for FrontRange and FrontRangeCLI.

Another solution would be to factor FrontRangeMCP out to a separate package, but this would require more work and maintenance. It's possible as a future option.